requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1
django-celery==3.1.17
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-cuser==2017.3.16
django-debug-toolbar==4.4.6
django-jazzmin==3.0.0
django-modeltranslation==0.19.9
django-supertools==0.3
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
kombu==5.4.1
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
